+++ config/routes.rb
@@ -5,6 +5,7 @@ require 'doc_auth_constraint'
 UUID_REGEX = /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/i.freeze
 
 Rails.application.routes.draw do
+  mount Auth::Engine, at: "/auth"
   namespace :api do
     post :graphql, to: 'graphql#create'