+++ engines/auth/lib/auth/exceptions.rb
@@ -1,0 +1,5 @@
+module Auth
+  module Exceptions
+    class BadRequestError < StandardError; end
+  end
+end